pyproject.toml
[project]
name = "pandora"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "milkymap", email = "ibrahima.elmokhtar@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "google-genai>=1.25.0",
    "openai>=1.95.1",
    "python-dotenv>=1.1.1",
    "pyzmq>=27.0.0",
]

[project.scripts]
pandora = "pandora:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
